+++ packages/smooth_app/lib/knowledge_panel/knowledge_panel_product_cards.dart
