+++ packages/smooth_app/lib/knowledge_panel/knowledge_panels/knowledge_panel_summary_card.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:openfoodfacts/model/KnowledgePanel.dart';
 import 'package:smooth_app/cards/data_cards/score_card.dart';
-import 'package:smooth_app/cards/product_cards/knowledge_panels/knowledge_panel_title_card.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
 import 'package:smooth_app/helpers/score_card_helper.dart';
+import 'package:smooth_app/knowledge_panel/knowledge_panels/knowledge_panel_title_card.dart';
 
 class KnowledgePanelSummaryCard extends StatelessWidget {
   const KnowledgePanelSummaryCard(this.knowledgePanel);
