+++ packages/smooth_app/lib/knowledge_panel/knowledge_panels/knowledge_panel_title_card.dart
